congestion_detector/requirements.txt
google-adk>=1.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
kafka-python>=2.0.2
redis>=4.5.0
joblib>=1.3.0
httpx>=0.24.0
python-dotenv>=1.0.0 